+++ CHANGELOG.md
@@ -19,6 +19,7 @@ For a full diff see [`2.0.0...main`][2.0.0...main].
 - Renamed `ExceptionInterface` to `Exception` ([#1125]), by [@localheinz]
 - Renamed `File` to `FilePath` ([#1126]), by [@localheinz]
 - Renamed `FileCouldNotBeParsed::fromFileAndParseError()` to `FileCouldNotBeParsed::atFilePathWithParseError()` ([#1127]), by [@localheinz]
+- Used `Source` instead of `string` when referencing source code ([#1130]), by [@localheinz]
 
 ## [`2.0.0`][2.0.0]
 
@@ -247,6 +248,7 @@ For a full diff see [`0.4.0...0.5.0`][0.4.0...0.5.0].
 [#1126]: https://github.com/ergebnis/classy/pull/1126
 [#1127]: https://github.com/ergebnis/classy/pull/1127
 [#1128]: https://github.com/ergebnis/classy/pull/1128
+[#1130]: https://github.com/ergebnis/classy/pull/1130
 
 [@ergebnis]: https://github.com/ergebnis
 [@localheinz]: https://github.com/localheinz